JottTokenizer: improve colon, fcHeader performance
This will be slightly faster and use less memory than the previous
implementation because it doesn't require us to create String objects,
instead statically passing what logically must be the token depending on
the branch we are in. It also removes the second conditional in the
final return statement by instead duplicating the return statement in
the two branches (colon / fcHeader).