JottTokenizer: Make sure that -1 is not passed to PushbackReader::unread
Passing -1 to PushbackReader::unread will result in a nasty overflow
edge case bug that we'll want to avoid.

The trouble can be anticipated by a close reading of the
PushbackReader::unread documentation (emphasis mine):

	Pushes back a single character by copying it to the front of the
	pushback buffer. After this method returns, the next character
	to be read will have the value **(char)c**.

The trouble is that unlike the integer types in Java, char is
*unsigned*, representing values from U+0000 to U+FFFF. So when we give
PushbackReader::unread the int value of -1, the next thing we get from
PushbackReader::read will be the int value of the char value of U+FFFF,
or, to put it another way, the result of (int)(char)-1. This means that
our EOF checks against -1 will fail on our next loop, if we encounter a
-1 from PushbackReader::read and then push it back.

I actually came across this bug because I was looking into removing the
`nc != -1` check I already had -- after all, it's a performance hit on
every case but one. This led to one of the test cases from the
professor failing, one that had an id/keyword at the end of the file,
*with no newline at the end*, as is common in most text files. We can
try this ourselves: on the commit before this one, replace the contents
of test_stuff.txt with a single ':' character -- be sure to include no
whitespace. This can be done on a UNIX-like system (like a Mac) with

	printf ':' > test_stuff.txt

tl;dr: We must ensure that we have this -1 check in any other branches
of the tokenizer where we push back a character with unread.